requirements.txt
Django==1.11.7
django-colorful==1.2
django-grappelli==2.10.1
django-simple-history==1.9.0
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pytz==2017.3